postgres_scraper/pyproject.toml
[project]
name = "postgres-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "loguru>=0.7.3",
    "openai>=1.68.2",
    "psycopg[binary]>=3.2.6",
    "pydantic>=2.10.6",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
